refactor: 앨범 분석 페이지 클라이언트 컴포넌트 파일 이동
