+++ src/app/(default)/album/analysis/[id]/_client-album-analysis.tsx
